+++ test/bigtest/network/factories/expenseClass.js
@@ -1,0 +1,7 @@
+import { Factory } from 'miragejs';
+import faker from 'faker';
+
+export default Factory.extend({
+  id: faker.random.uuid,
+  name: faker.finance.accountName,
+});